+++ test/core/req_len_test.dart
@@ -5,8 +5,8 @@ void main() {
   group('ReqLen —', () {
     const blank = 'cannot be left blank';
     const diff = 'value length diff';
-    const less = 'the value is too small';
-    const great = 'the value is too large';
+    const short = 'the value is too small';
+    const long = 'the value is too large';
     group('assertions and error:', () {
       test('invalid length', () {
         expect(() => ReqLen(-1), throwsA(isA<AssertionError>()));
@@ -86,32 +86,32 @@ void main() {
       });
 
       group('Len.min', () {
-        final minFive = ReqLen.min(5, blank: blank, less: less);
+        final minFive = ReqLen.min(5, blank: blank, short: short);
         test('String', () {
           expect(minFive(null), blank);
           expect(minFive(''), blank);
-          expect(minFive('1'), less);
+          expect(minFive('1'), short);
           expect(minFive('12345'), null);
           expect(minFive('123456'), null);
         });
         test('num', () {
           expect(minFive(null), blank);
-          expect(minFive(1), less);
-          expect(minFive(01234), less);
+          expect(minFive(1), short);
+          expect(minFive(01234), short);
           expect(minFive(12345), null);
           expect(minFive(123456), null);
         });
         test('Iterable', () {
           expect(minFive(null), blank);
           expect(minFive(<String>[]), blank);
-          expect(minFive(['1']), less);
+          expect(minFive(['1']), short);
           expect(minFive([1, 2, 3, 4, 5]), null);
           expect(minFive([1, 2, 3, 4, 5, 6]), null);
         });
         test('Map', () {
           expect(minFive(null), blank);
           expect(minFive(<int, int>{}), blank);
-          expect(minFive({1: 'a'}), less);
+          expect(minFive({1: 'a'}), short);
           expect(minFive({1: 'a', 2: 'b', 3: 'c', 4: 'd', 5: 'e'}), null);
           expect(
             minFive({1: 'a', 2: 'b', 3: 'c', 4: 'd', 5: 'e', 6: 'f'}),
@@ -121,26 +121,26 @@ void main() {
       });
 
       group('Len.max', () {
-        final maxFive = Len.max(5, great: great);
+        final maxFive = Len.max(5, long: long);
         test('String', () {
           expect(maxFive(null), null);
           expect(maxFive(''), null);
           expect(maxFive('a'), null);
           expect(maxFive('12345'), null);
-          expect(maxFive('123456'), great);
+          expect(maxFive('123456'), long);
         });
         test('num', () {
           expect(maxFive(null), null);
           expect(maxFive(1), null);
           expect(maxFive(12345), null);
-          expect(maxFive(123456), great);
+          expect(maxFive(123456), long);
         });
         test('Iterable', () {
           expect(maxFive(null), null);
           expect(maxFive(<String>[]), null);
           expect(maxFive(['a']), null);
           expect(maxFive([1, 2, 3, 4, 5]), null);
-          expect(maxFive([1, 2, 3, 4, 5, 6]), great);
+          expect(maxFive([1, 2, 3, 4, 5, 6]), long);
         });
         test('Map', () {
           expect(maxFive(null), null);
@@ -149,33 +149,33 @@ void main() {
           expect(maxFive({1: 'a', 2: 'b', 3: 'c', 4: 'd', 5: 'e'}), null);
           expect(
             maxFive({1: 'a', 2: 'b', 3: 'c', 4: 'd', 5: 'e', 6: 'f'}),
-            great,
+            long,
           );
         });
       });
 
       group('Len.range', () {
         final rangeOneFive =
-            ReqLen.range(1, 5, blank: blank, less: less, great: great);
+            ReqLen.range(1, 5, blank: blank, short: short, long: long);
         test('String', () {
           expect(rangeOneFive(null), blank);
           expect(rangeOneFive(''), blank);
           expect(rangeOneFive('a'), null);
           expect(rangeOneFive(12345), null);
-          expect(rangeOneFive(123456), great);
+          expect(rangeOneFive(123456), long);
         });
         test('num', () {
           expect(rangeOneFive(null), blank);
           expect(rangeOneFive(0), null);
           expect(rangeOneFive(12345), null);
-          expect(rangeOneFive(123456), great);
+          expect(rangeOneFive(123456), long);
         });
         test('Iterable', () {
           expect(rangeOneFive(null), blank);
           expect(rangeOneFive(<String>[]), blank);
           expect(rangeOneFive(['a']), null);
           expect(rangeOneFive([1, 2, 3, 4, 5]), null);
-          expect(rangeOneFive([1, 2, 3, 4, 5, 6]), great);
+          expect(rangeOneFive([1, 2, 3, 4, 5, 6]), long);
         });
         test('Map', () {
           expect(rangeOneFive(null), blank);
@@ -184,7 +184,7 @@ void main() {
           expect(rangeOneFive({1: 'a', 2: 'b', 3: 'c', 4: 'd', 5: 'e'}), null);
           expect(
             rangeOneFive({1: 'a', 2: 'b', 3: 'c', 4: 'd', 5: 'e', 6: 'f'}),
-            great,
+            long,
           );
         });
       });