+++ test/numeric/hex_test.dart
@@ -5,16 +5,16 @@ void main() {
   group('Hex —', () {
     const error = 'the data must only contain hexadecimal digits';
     const diff = 'length error';
-    const less = 'the length is too short';
-    const great = 'the length is too long';
+    const short = 'the length is too short';
+    const long = 'the length is too long';
     const len = 2;
     const min = 3;
     const max = 8;
     final dig = Hex(mal: error);
     final digLen = Hex.len(len, mal: error, diff: diff);
-    final digMin = Hex.min(min, mal: error, less: less);
-    final digMax = Hex.max(max, mal: error, great: great);
-    final digRange = Hex.range(min, max, mal: error, less: less, great: great);
+    final digMin = Hex.min(min, mal: error, short: short);
+    final digMax = Hex.max(max, mal: error, long: long);
+    final digRange = Hex.range(min, max, mal: error, short: short, long: long);
     test('null', () {
       expect(dig(null), null);
       expect(digLen(null), null);
@@ -33,9 +33,9 @@ void main() {
       const twoHex = 'FF';
       expect(dig(twoHex), null);
       expect(digLen(twoHex), null);
-      expect(digMin(twoHex), less);
+      expect(digMin(twoHex), short);
       expect(digMax(twoHex), null);
-      expect(digRange(twoHex), less);
+      expect(digRange(twoHex), short);
     });
     test('4 digits', () {
       const fourHexs = 'fe98';
@@ -58,8 +58,8 @@ void main() {
       expect(dig(tenHex), null);
       expect(digLen(tenHex), diff);
       expect(digMin(tenHex), null);
-      expect(digMax(tenHex), great);
-      expect(digRange(tenHex), great);
+      expect(digMax(tenHex), long);
+      expect(digRange(tenHex), long);
     });
     test('the last char is a non-hex-digit', () {
       const invalid = '0123456789ABCDEFX';